% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anova_analysis}
\alias{anova_analysis}
\title{Compute anova analysis for each protein to compare each method altogether}
\usage{
anova_analysis(log2)
}
\arguments{
\item{log2}{protein abundance matrix (log2 normalized)}
}
\value{
anova output matrix
}
\description{
Compute anova analysis for each protein to compare each method altogether
}
\examples{
anova_analysis(log2)
}
