% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{col_sorted}
\alias{col_sorted}
\title{Format your table for others functions}
\usage{
col_sorted(data)
}
\arguments{
\item{data.frame}{to be reformated}
}
\value{
a long data.frame (melted)
}
\description{
Format your table for others functions
}
\examples{
data("ttest")
col_sorted(ttest)
}
