% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{column_maker}
\alias{column_maker}
\title{column parser for analysis}
\usage{
column_maker(df, colname2check, newcolname)
}
\arguments{
\item{df}{the dataframe of interest}

\item{colname2check}{the column of interest}

\item{newcolname}{the returned column}
}
\value{
a new dataframe with a new column
}
\description{
column parser for analysis
}
\examples{
column_maker(datamined, "Glycosite", "Glycological")
}
