% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{combineMembraneColumns}
\alias{combineMembraneColumns}
\title{Combine all the membrane terms from the Subcell Table}
\usage{
combineMembraneColumns(df)
}
\arguments{
\item{df}{output of table_maker_from_strings(data$Subcell, data)}
}
\value{
the same table but with all the Membrane columns grouped in one
}
\description{
Combine all the membrane terms from the Subcell Table
}
\examples{
combineMembraneColumns(subcell)
}
