% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_meandiff_methods}
\alias{compute_meandiff_methods}
\title{Compare the mean of one method against all others per protein}
\usage{
compute_meandiff_methods(df)
}
\arguments{
\item{df}{log2 protein abundance matrix}

\item{padj}{p.adjustment method}
}
\value{
mean differences matrix
}
\description{
Compare the mean of one method against all others per protein
}
\examples{
compute_meandiff_methods(log2)
}
