% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_ttest}
\alias{compute_ttest}
\title{Compute T-Test of one method against all others per protein}
\usage{
compute_ttest(df, padj = "fdr")
}
\arguments{
\item{df}{log2 protein abundance matrix}

\item{padj}{p.adjustment method}
}
\value{
ttest matrix of each method
}
\description{
Compute T-Test of one method against all others per protein
}
\examples{
compute_ttest(log2)
}
