% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dotplot_maker}
\alias{dotplot_maker}
\title{Enrichment plot of your data}
\usage{
dotplot_maker(
  ftest,
  namefile = "FisherDotPlot_Enrichment.png",
  path = "",
  width = 25,
  height = 6,
  saveplot = FALSE
)
}
\arguments{
\item{ftest}{The output of fishertest()}

\item{path}{will define directory to store output. by default, files are saved within wd}
}
\value{
A plot of the fisher test enrichment
}
\description{
Enrichment plot of your data
}
\examples{
data("ftest")
dotplot_maker(fisher_subcell, width=25, height=10)
}
