% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dotratio}
\alias{dotratio}
\title{Create a simple dotplot around the max and min values per method}
\usage{
dotratio(
  logicaltable,
  log2,
  mean_computing = TRUE,
  saveplot = FALSE,
  namefile = "DotPlot.png",
  path = "",
  width = 25,
  height = 6
)
}
\arguments{
\item{logicaltable}{logical table of the mined properties}

\item{log2}{log2 protein abundances matrix}

\item{mean_computing}{True/False; if True, you will get average values to get one color for the dot. If False, applies a gradient within each dot based on values.}

\item{saveplot}{True/False to save the plot in the desired path}

\item{namefile}{name of the output file if saved}

\item{path}{path of the output file if saved}

\item{width}{width of the output file if saved}

\item{height}{height of the output file if saved}
}
\value{
dotplot around the max and min values per method
}
\description{
Create a simple dotplot around the max and min values per method
}
\examples{
dotratio(subcell, log2)
}
