% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{extract_glyco_string}
\alias{extract_glyco_string}
\title{Transform glycosylation enrichment as a logisitc table enumrating O & N glycosylation}
\usage{
extract_glyco_string(table, colname)
}
\arguments{
\item{table}{the dataframe containing the glycosylation site from neXtProt}

\item{colname}{the colname of interest}
}
\value{
a new dataframe, with 2 logistic column O_linked and N_linked
}
\description{
Transform glycosylation enrichment as a logisitc table enumrating O & N glycosylation
}
\examples{
extract_glyco_string(datamined, "Glycosite")
}
