% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fishertest}
\alias{fishertest}
\title{Makes a Fisher Exact Test to get enrichment of your logical columns}
\usage{
fishertest(df_mined, ttest_logical, padj = "none")
}
\arguments{
\item{ttest_logical}{A logical data-frame of the ouput of mypairwise_t().}

\item{padj}{Correction method for your p.value. See ?p.adjust}

\item{df_mined_logical}{Is a dataframe of your data.set but with biophysical properties}
}
\value{
A fisher table with the numbers that were used to make the f.test, and the p-value and p-adjust.
}
\description{
Makes a Fisher Exact Test to get enrichment of your logical columns
}
\examples{
data("ttest")
# taille du points: nombre de gene T_T
# couleur du point: p.value
data("big")
fishertest(reformat_table(big), reformat_table(ttest), "fdr")
}
