% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_annotationsnumber}
\alias{get_nextprot_annotationsnumber}
\title{Title General function to mine integer annotations (number of TM domains, catalytic activity...)}
\usage{
get_nextprot_annotationsnumber(protname, name_of_annotation)
}
\arguments{
\item{protname}{UniProtID}

\item{name_of_annotation}{name of the string (separated with "-")}
}
\value{
the number of annotations
}
\description{
Title General function to mine integer annotations (number of TM domains, catalytic activity...)
}
\examples{
get_nextprot_annotationsnumber("O15118", "transmembrane-region")
}
