% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_coiledcoil}
\alias{get_nextprot_coiledcoil}
\title{Get number of coiled-coil of the protein}
\usage{
get_nextprot_coiledcoil(protname)
}
\arguments{
\item{protname}{UniProtID}
}
\value{
number of coiled-coil
}
\description{
Get number of coiled-coil of the protein
}
\examples{
get_nextprot_coiledcoil("Q99996")
}
