% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_glycosite}
\alias{get_nextprot_glycosite}
\title{get glycosylation for a given protein}
\usage{
get_nextprot_glycosite(protname)
}
\arguments{
\item{protname}{UniProtID}
}
\value{
the type of glycosylation in strings or vector if few
}
\description{
get glycosylation for a given protein
}
\examples{
get_nextprot_glycosite("Q8N2K0")
}
