% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_lipidationsite}
\alias{get_nextprot_lipidationsite}
\title{Title Get number of lipidation site of the protein}
\usage{
get_nextprot_lipidationsite(protname)
}
\arguments{
\item{protname}{UniProtID}
}
\value{
number of lipidation site
}
\description{
Title Get number of lipidation site of the protein
}
\examples{
get_nextprot_lipidationsite("Q14254")
}
