% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_mass}
\alias{get_nextprot_mass}
\title{get protein mass}
\usage{
get_nextprot_mass(protname)
}
\arguments{
\item{protname}{UniProtID}
}
\value{
Mass (Da)
}
\description{
get protein mass
}
\examples{
get_nextprot_mass("P01023")
}
