% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_PI}
\alias{get_nextprot_PI}
\title{get isoelectric point}
\usage{
get_nextprot_PI(protname)
}
\arguments{
\item{SwissProtID}{}
}
\value{
Isoelectric Point
}
\description{
get isoelectric point
}
\examples{
get_nextprot_PI("P01023")
}
