% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_subcell}
\alias{get_nextprot_subcell}
\title{Get subcellular locations}
\usage{
get_nextprot_subcell(protname)
}
\arguments{
\item{protname}{}
}
\value{
subcellular location(s) in strings
}
\description{
Get subcellular locations
}
\examples{
get_nextprot_subcell("P01023")
}
