% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_transmembrane}
\alias{get_nextprot_transmembrane}
\title{get number of TM domains}
\usage{
get_nextprot_transmembrane(protname)
}
\arguments{
\item{protein}{name}
}
\value{
number of TM domains
}
\description{
get number of TM domains
}
\examples{
get_nextprot_transmembrane("Q9NRK6")
}
