% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_nextprot_zincfinger}
\alias{get_nextprot_zincfinger}
\title{Title Get number of Zinc-Finger region of the protein}
\usage{
get_nextprot_zincfinger(protname)
}
\arguments{
\item{protname}{UniProtID}
}
\value{
number of zinc-finger region
}
\description{
Title Get number of Zinc-Finger region of the protein
}
\examples{
get_nextprot_zincfinger("O14686")
}
