% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{get_swissprot_id}
\alias{get_swissprot_id}
\title{get canonical SwissProt ID}
\usage{
get_swissprot_id(gene_name, organism = 9606, reviewed = "true")
}
\arguments{
\item{gene_name}{gene name to transform}

\item{organism}{taxonomy ID}

\item{reviewed}{(True/False)}
}
\value{
Canonical SwissProtID
}
\description{
get canonical SwissProt ID
}
\examples{
get_swissprot_id("A2M")
}
