% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggpairs_proteomics}
\alias{ggpairs_proteomics}
\title{Title}
\usage{
ggpairs_proteomics(log2, datamined, column2highlight)
}
\arguments{
\item{log2}{log2 protein abundances table}

\item{datamined}{datamined table}

\item{column2highlight}{factor or logistic column to highlight plots}
}
\value{
exploratory plots from ggpairs
}
\description{
Title
}
\examples{
ggpairs_proteomics(log2, datamined, "Groups.of.TM")
}
