% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap_maker}
\alias{heatmap_maker}
\title{Title}
\usage{
heatmap_maker(num_df, scale_data = TRUE)
}
\arguments{
\item{num_df}{A df with numeric values}

\item{scale_data}{A logical parameter scaling your data or not. Note that data should be scaled.}
}
\value{
Heatmap of your methods
}
\description{
Title
}
\examples{
data("log2")
heatmap_maker(log2[,-1])
}
