% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lm_analysis}
\alias{lm_analysis}
\title{Compute linear models for each protein to compare each method}
\usage{
lm_analysis(log2)
}
\arguments{
\item{log2}{protein abundance table}
}
\value{
a matrix summarizing results of lm
}
\description{
Compute linear models for each protein to compare each method
}
\examples{
lm_analysis(log2)
}
