% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean_diff_methods}
\alias{mean_diff_methods}
\title{Compare the mean of one method against all others per protein}
\usage{
mean_diff_methods(df, padj = "fdr")
}
\arguments{
\item{df}{log2 protein abundance matrix}

\item{padj}{p.adjustment method}
}
\value{
mean differences matrix
}
\description{
Compare the mean of one method against all others per protein
}
\examples{
mean_diff_methods(log2)
}
