% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mean_function}
\alias{mean_function}
\title{This function makes the mean of your replicates for each methods}
\usage{
mean_function(df)
}
\arguments{
\item{df}{with protein abundances}
}
\value{
A data.frame of the mean of each method
}
\description{
This function makes the mean of your replicates for each methods
}
\examples{
data("log2")
mean_function(log2)
}
