% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meandiffploting}
\alias{meandiffploting}
\title{Create a Plot of the difference of the means after using T-Test and MeanDiff}
\usage{
meandiffploting(
  tablettest,
  tablemean,
  minedproperties,
  saveplot = FALSE,
  namefile = "MeanDifferenceDotPlot",
  path = "",
  width = 10,
  height = 10
)
}
\arguments{
\item{tablettest}{output of compute_ttest()}

\item{tablemean}{output of mean_diff_methods()}

\item{minedproperties}{logistic table of mined properties}

\item{saveplot}{True/False to save the plot in the desired path}

\item{namefile}{name of the output file if saved}

\item{path}{path of the output file if saved}

\item{width}{width of the output file if saved}

\item{height}{height of the output file if saved}
}
\value{
plot
}
\description{
Create a Plot of the difference of the means after using T-Test and MeanDiff
}
\examples{
meandifftesting(meantt, meandiff, glycolink)
}
