% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myfoldchange_unpaired}
\alias{myfoldchange_unpaired}
\title{Compute fold-change for each method vs all other methods}
\usage{
myfoldchange_unpaired(log2)
}
\arguments{
\item{log2}{protein abundance table (log2 normalized)}
}
\value{
foldchange matrix for each method
}
\description{
Compute fold-change for each method vs all other methods
}
\examples{
myfoldchange_unpaired(log2)
}
