% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myMDSplot}
\alias{myMDSplot}
\title{Creates MDS Plot from Log2 Protein Abundances to show how the different methods are close to each other}
\usage{
myMDSplot(log2, rows = TRUE)
}
\arguments{
\item{log2}{your first table with log2 normalized data of protein abundances, with a "name" column of your identifiers}

\item{rows}{If True, rownames will take the first column and delete it from the data frame}
}
\value{
A MDS Plot of your methods
}
\description{
Creates MDS Plot from Log2 Protein Abundances to show how the different methods are close to each other
}
\examples{
data("log2")
myMDSplot(log2)
}
