% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mypairwise_t}
\alias{mypairwise_t}
\title{Automatize pairwise t.test of your data.}
\usage{
mypairwise_t(df, padj = "fdr")
}
\arguments{
\item{df}{A data.frame with protein abundances, with a "name" column of your identifiers}

\item{padj}{A string parameter defining the correction methods. See ?p.adjust}
}
\value{
A table of pairwise t.test of 2 times your amount of methods
}
\description{
Automatize pairwise t.test of your data.
}
\examples{
data("log2")
mypairwise_t(log2)
}
