% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myvenndiagram}
\alias{myvenndiagram}
\title{Automatize and Save VennDigram of the result of the mypairwise_t() above}
\usage{
myvenndiagram(output_ttest, path = "examples/", saveplot = FALSE)
}
\arguments{
\item{path}{will define directory to store output. by default, files are saved within wd}

\item{output}{of mypairwise_t()}
}
\value{
Nothing is returned, but x plots will be stored, x being the amount of methods in your dataset
}
\description{
Automatize and Save VennDigram of the result of the mypairwise_t() above
}
\examples{
data("ttest")
myvenndiagram(ttest)
}
