% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{proteinid_parser}
\alias{proteinid_parser}
\title{Choose which proteins to analyze from MaxQuant format}
\usage{
proteinid_parser(df, colname, sep = ";", KeepAllProteins = FALSE)
}
\arguments{
\item{df}{a df with a dataframe containing one or more protein per row}

\item{colname}{the column of interest in string}

\item{sep}{the separator for the proteins}

\item{KeepAllProteins}{logical. If True, all the proteins will be added to the dataset.}
}
\value{
a dataframe with only one protein per row
}
\description{
Choose which proteins to analyze from MaxQuant format
}
\examples{
proteinid_parser(Log2_Extraction_comparison_Sep22, "Majority.protein.IDs")
}
