% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reformat_table}
\alias{reformat_table}
\title{This function ensure to get the right input for fishertest()}
\usage{
reformat_table(df, padj = "fdr", treshold = 0.05)
}
\arguments{
\item{df}{A Dataframe that needs to be logical. Numeric value like p-values will be transformed as logical depending to a certain treshold. Integer values will be taken as default (0=false, else=true)}

\item{padj}{Correction method. See p.adjust()}

\item{treshold}{The p-value treshold.}
}
\value{
a data.frame with only logical variable instead of integer and numeric variables.
}
\description{
This function ensure to get the right input for fishertest()
}
\examples{
data("ttest")
reformat_table(ttest)
}
