% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{subcell_table_maker}
\alias{subcell_table_maker}
\title{Transform list of subcell locations into a wide logistic table.}
\usage{
subcell_table_maker(df)
}
\arguments{
\item{df}{}
}
\value{
logistic table
}
\description{
Transform list of subcell locations into a wide logistic table.
}
\examples{
subcell_table_maker(dataf)
}
