% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{table_maker_from_strings}
\alias{table_maker_from_strings}
\title{Transform list of subcell locations into a wide logistic table.}
\usage{
table_maker_from_strings(df, colnamestring)
}
\arguments{
\item{df}{containing subcell strings after get_nextprot_subcell() with colname "SubcellStrings"}
}
\value{
logistic table of the mined strings
}
\description{
Transform list of subcell locations into a wide logistic table.
}
\examples{
table_maker_from_strings(datamined, "SubcellStrings")
}
