% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{uniprot_assembling}
\alias{uniprot_assembling}
\title{Create a table traducing gene name into canonical UniProtID}
\usage{
uniprot_assembling(df)
}
\arguments{
\item{df}{containing gene names in first column}
}
\value{
df containing UniProtID
}
\description{
Create a table traducing gene name into canonical UniProtID
}
\examples{
uniprot_assembling(big)
}
