% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{volcano_all}
\alias{volcano_all}
\title{Compare your volcano plots together}
\usage{
volcano_all(
  ttest,
  foldc,
  datamined,
  binarycol,
  threshold = 5,
  namefile = "Volcano_",
  path = "",
  width = 40,
  height = 6,
  saveplot = FALSE
)
}
\arguments{
\item{ttest}{output of mypairwise_t, with a "name" column of your identifiers}

\item{foldc}{output of myfoldchange, with a "name" column of your identifiers}

\item{datamined}{dataframe with a logical variable to highlight, with a "name" column of your identifiers}

\item{binarycol}{the variable to highlight}

\item{threshold}{should depend on the p.value adjustment, you can adapt it to your data}

\item{path}{will define directory to store output. by default, files are saved within wd}
}
\value{
Volcanoplots for each pair of methods
}
\description{
Compare your volcano plots together
}
\examples{
data("ttest")
data("foldc")
volcano_all(ttest, foldc, datamined, "Groups.of.TM")
}
