% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{volcano_maker}
\alias{volcano_maker}
\title{Automatize VolcanoPlots of your dataset}
\usage{
volcano_maker(
  ttest,
  foldc,
  datamined,
  binarycol,
  threshold = 5,
  namefile = "Volcano",
  path = "examples/",
  width = 10,
  height = 10,
  saveplot = FALSE
)
}
\arguments{
\item{ttest}{A data.frame of p.values (Output of mypairwise_t())}

\item{foldc}{A data.frame of fold changes (Output of myfoldchange())}

\item{datamined}{A reformated data.frame of mined data (reformat_table(datamined))}

\item{binarycol}{A String parameter of the column you wish to highlight}

\item{threshold}{should depend on the p.value adjustment, you can adapt it to your data}

\item{path}{will define directory to store output. by default, files are saved within working directory}

\item{saveplot}{boolean parameter to whether save or not the plot}
}
\value{
Saves X plots of your pairs of methods, X being 2 times the amount of methods
}
\description{
Automatize VolcanoPlots of your dataset
}
\examples{
data("ttest")
data("foldc")
data("datamined")
volcano_maker(ttest, foldc, reformat_table(datamined), "Groups.of.TM")
}
