% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{datamining_annotationsnumber}
\alias{datamining_annotationsnumber}
\title{Incorporate a new column from the datamining function "get_nextprot_annotationsnumber")}
\usage{
datamining_annotationsnumber(df, fun, annotation)
}
\arguments{
\item{df}{containing "UniProtID" as colname}

\item{fun}{name of the function to use}

\item{annotation}{name of the annotation parameter of the function to use}

\item{colname}{in strings of the datamined column}
}
\value{
df containing the new column
}
\description{
Incorporate a new column from the datamining function "get_nextprot_annotationsnumber")
}
\examples{
datamining_annotationsnumber(head(datamined), "catalytic_activity", get_nextprot_annotationsnumber, "catalytic-activity")
}
