% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dotplot_by_methods}
\alias{dotplot_by_methods}
\title{create a dotplot colored by Zscore and sized by the percent of abundance}
\usage{
dotplot_by_methods(
  logistictable,
  log2,
  saveplot = FALSE,
  namefile = "DotPlot.png",
  path = "",
  width = 25,
  height = 6
)
}
\arguments{
\item{logistictable}{a dataframe with mined data contained as logistic variables; with protein/gene name as "name"}

\item{log2}{a dataframe with protein abundances; with protein/gene name as "name"}

\item{saveplot}{if True, the plot will be saved}

\item{namefile}{the name of the plot to save if saveplot=True}

\item{path}{the path of the output to save if saveplot=True}

\item{width}{the width of the plot to save if saveplot=True}

\item{height}{the height of the plot to save if saveplot=True}
}
\value{
dotplot; saved or not
}
\description{
create a dotplot colored by Zscore and sized by the percent of abundance
}
\examples{
dotplot_by_methods(logistic_mined_data, log2)
}
