% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggpairs_proteomics}
\alias{ggpairs_proteomics}
\title{ggpairs plot with extra highlight on a specific column}
\usage{
ggpairs_proteomics(log2, datamined, column2highlight)
}
\arguments{
\item{log2}{log2 protein abundances table; with "name" column which is the protein/gene identifier}

\item{datamined}{datamined table, with the exact same column "name"}

\item{column2highlight}{factor or logical column to highlight plots}
}
\value{
exploratory plots from ggpairs
}
\description{
ggpairs plot with extra highlight on a specific column
}
\examples{
ggpairs_proteomics(log2, datamined, "Groups.of.TM")
}
