% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myfoldchange}
\alias{myfoldchange}
\title{Makes foldchange around your methods}
\usage{
myfoldchange(log2)
}
\arguments{
\item{df}{with protein abundances in mean. removing replicates can be done with mean_function(). gene name should be named "name" if you want to keep it as rowname.}
}
\value{
A data.frame of the different fold.change values.
}
\description{
Makes foldchange around your methods
}
\examples{
data("log2")
myfoldchange(log2)
}
