% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myfoldchange}
\alias{myfoldchange}
\title{Makes foldchange around your methods}
\usage{
myfoldchange(df)
}
\arguments{
\item{long_df}{Melted data.frame with replicates_factor()}
}
\value{
A data.frame of the different fold.change values.
}
\description{
Makes foldchange around your methods
}
\examples{
data("log2")
myfoldchange(mean_function(replicates_factor(log2)))
}
