% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mypairwise_t}
\alias{mypairwise_t}
\title{Automatize pairwise t.test of your data.}
\usage{
mypairwise_t(long_df, padj = "none")
}
\arguments{
\item{long_df}{A data.frame that is melted. You can use replicates_factor() below}

\item{padj}{A string parameter defining the correction methods. See p.adjust()}
}
\value{
A table of pairwise t.test of 2 times your amount of methods
}
\description{
Automatize pairwise t.test of your data.
}
\examples{
data("log2")
mypairwise_t(replicates_factor(log2))
}
