% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replicates_factor}
\alias{replicates_factor}
\title{Used melt() and separate() to reformat your table into a long one, which will seperate your replicates as a new variable.}
\usage{
replicates_factor(df)
}
\arguments{
\item{df}{Unmelted df (log2)}
}
\value{
A long and reformated data.frame
}
\description{
Used melt() and separate() to reformat your table into a long one, which will seperate your replicates as a new variable.
}
\examples{
data("log2")
replicates_factor(log2)
}
