% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_functions.R
\name{table_maker_from_strings}
\alias{table_maker_from_strings}
\title{Transform list of subcell locations into a wide logistic table.}
\usage{
table_maker_from_strings(csv_column, df = NULL)
}
\arguments{
\item{csv_column}{the specific column with the comma separated string}

\item{df}{facultative data.frame to retrieve rownames if needed}
}
\value{
logistic table of the mined strings
}
\description{
Transform list of subcell locations into a wide logistic table.
}
\examples{
table_maker_from_strings(newdata$Subcell, newdata)
}
