% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{volcano_all}
\alias{volcano_all}
\title{Compare your volcano plots together}
\usage{
volcano_all(ttest, foldc, big, binarycol, threshold = 5, path = "examples/")
}
\arguments{
\item{ttest}{output of mypairwise_t}

\item{foldc}{output of myfoldchange}

\item{big}{dataframe with a logical variable to highlight}

\item{binarycol}{the variable to highlight}

\item{threshold}{should depend on the p.value adjustment, you can adapt it to your data}

\item{path}{will define directory to store output. by default, files are saved within wd}
}
\value{
Volcanoplots for each pair of methods
}
\description{
Compare your volcano plots together
}
\examples{
data("ttest")
data("foldc")
volcano_all(ttest, foldc, big, "Groups.of.TM")
# to visualize each variable in a for loop, you can use colnames():
# for (colname in colnames(big)[-1]) {volcano_all(ttest, foldc, big, colname)}
}
