% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{volcano_maker}
\alias{volcano_maker}
\title{Automatize VolcanoPlots of your dataset}
\usage{
volcano_maker(
  dfpval,
  dffoldchange,
  binarydf,
  binarycol,
  threshold = 5,
  path = ""
)
}
\arguments{
\item{dfpval}{A data.frame of p.values (Output of mypairwise_t())}

\item{dffoldchange}{A data.frame of fold changes (Output of myfoldchange())}

\item{binarydf}{A reformated data.frame of mined data (reformat_table(big))}

\item{binarycol}{A String parameter of the column you wish to investigate}

\item{threshold}{should depend on the p.value adjustment, you can adapt it to your data}

\item{path}{will define directory to store output. by default, files are saved within wd}
}
\value{
Saves X plots of your pairs of methods, X being 2 times the amount of methods
}
\description{
Automatize VolcanoPlots of your dataset
}
\examples{
data("ttest")
data("foldc")
data("big")
volcano_maker(ttest, foldc, reformat_table(big), "Groups.of.TM")
}
